Refine timing output in with_procmaps function
- Updated the time measurement in the with_procmaps function to display the duration with four decimal places for improved precision.
- This change enhances the clarity of performance monitoring in the XOM library.

+++ libxom.c
@@ -102,7 +102,7 @@ static int with_procmaps(callback_t callback, void *arg) {
     }
 
     fclose(file);
-    printf("Time taken: %ld seconds\n", (clock() - start_time) / CLOCKS_PER_SEC);
+    printf("Time taken: %.4f seconds\n", (double)(clock() - start_time) / CLOCKS_PER_SEC);
     return ret;
 }
